<==== Mercurial data ====>
Branch: x_CLOUD_222_nat_traversal

+++ unit_tests/cloud_connectivity_ut/src/cloud_stream_socket_ut.cpp
@@ -64,7 +64,7 @@ TEST_F(CloudStreamSocketTest, simple)
         tempHostName,
         serverAddress);
 
-    for (int i = 0; i < repeatCount; ++i)
+    for (size_t i = 0; i < repeatCount; ++i)
     {
         //connecting with CloudStreamSocket to the local server
         CloudStreamSocket cloudSocket;
@@ -203,7 +203,7 @@ TEST_F(CloudStreamSocketTest, cancellation)
         serverAddress);
 
     //cancelling connect
-    for (int i = 0; i < repeatCount; ++i)
+    for (size_t i = 0; i < repeatCount; ++i)
     {
         //connecting with CloudStreamSocket to the local server
         CloudStreamSocket cloudSocket;
@@ -214,7 +214,7 @@ TEST_F(CloudStreamSocketTest, cancellation)
     }
 
     //cancelling read
-    for (int i = 0; i < repeatCount; ++i)
+    for (size_t i = 0; i < repeatCount; ++i)
     {
         //connecting with CloudStreamSocket to the local server
         CloudStreamSocket cloudSocket;
@@ -230,7 +230,7 @@ TEST_F(CloudStreamSocketTest, cancellation)
     }
 
     //cancelling send
-    for (int i = 0; i < repeatCount; ++i)
+    for (size_t i = 0; i < repeatCount; ++i)
     {
         //connecting with CloudStreamSocket to the local server
         CloudStreamSocket cloudSocket;
